Fix 64-bit build.

+++ xen/include/asm-x86/mach-es7000/mach_mpparse.h
@@ -12,7 +12,7 @@ static inline int mps_oem_check(struct mp_config_table *mpc, char *oem,
 {
 	if (mpc->mpc_oemptr) {
 		struct mp_config_oemtable *oem_table = 
-			(struct mp_config_oemtable *)mpc->mpc_oemptr;
+			(struct mp_config_oemtable *)(long)mpc->mpc_oemptr;
 		if (!strncmp(oem, "UNISYS", 6))
 			return parse_unisys_oem((char *)oem_table);
 	}
